+++ tables.js
@@ -764,3 +764,1538 @@ var months = [
 "Morgat"
 ]
 
+var occupational_skills = {
+    "Alchemist": [
+        "Alchemy/4",
+        "Mathematics/3",
+        "Herblore/3",
+        "Mining/2",
+        "Physician/2",
+        "Script/5"
+    ],
+    "Animal Trainer": [
+        "Animalcraft/4",
+        "Hidework/2",
+        "RIDING/2"
+    ],
+    "Apothecary": [
+        "Herblore/4",
+        "Alchemy/2",
+        "Mathematics/2",
+        "Physician/2",
+        "Script/5"
+    ],
+    "Astrologer": [
+        "Astrology/4",
+        "Mathematics/3",
+        "Drawing/3",
+        "Tarotry/2",
+        "Script/5"
+    ],
+    "Baliff": [
+        "Law/3",
+        "Agriculture/2",
+        "INITIATIVE/6",
+        "Foraging/4",
+        "Survival/4",
+        "Physician/2",
+        "Heraldry/2",
+        "Weaponcraft/2",
+        "RIDING/6",
+        "Polearm(Lance)/6",
+        "Sword(Broadsword)/5",
+        "Axe(Handaxe)/5",
+        "Dagger/5",
+        "Shield(Knight)/6",
+        "Dancing/3"
+    ],
+    "Beggar": [
+        "RHETORIC/5",
+        "INTRIGUE/4",
+        "Dagger/4"
+    ],
+    "Cartographer/Artist": [
+        "Drawing/4",
+        "Mathematics/2",
+        "Script/5"
+    ],
+    "Chandler": [
+        "Brewing/3",
+        "Woodcraft/3",
+        "Hidework/3",
+        "Alchemy/2",
+        "Metalcraft/2",
+        "Script/5"
+    ],
+    "Charcoaler": [
+        "Timbercraft/3",
+        "Survival/3",
+        "Woodcraft/2",
+        "Metalcraft/2"
+    ],
+    "Chieftan": [
+        "Tracking/4",
+        "Bow/3",
+        "Spear/4",
+        "Dagger/4",
+        "STEALTH/4",
+        "Foraging/3",
+        "Survival/3",
+        "Hidework/2",
+        "Fletching/2",
+        "RHETORIC/4",
+        "INTRIGUE/4"
+    ],
+    "Clothier": [
+        "Textilecraft/4",
+        "Hidework/3",
+        "Jewelcraft/3",
+        "Script/5"
+    ],
+    "Cook/Servant": [
+        "INTRIGUE/4",
+        "RHETORIC/4",
+        "Cookery/4",
+        "Textilecraft/3",
+        "Herblore/2"
+    ],
+    "Courtesan": [
+        "Lovecraft/5",
+        "INTRIGUE/5",
+        "SINGING/5",
+        "Musician/3",
+        "Dancing/3",
+        "Script/5"
+    ],
+    "Embalmer": [
+        "Embalming/4",
+        "Woodcraft/2",
+        "Alchemy/2",
+        "Perfumery/2",
+        "Script/5"
+    ],
+    "Farmer": [
+        "Agriculture/4",
+        "Weatherlore/4",
+        "Animalcraft/3"
+    ],
+    "Fisherman": [
+        "Fishing/5",
+        "Seamanship/4",
+        "Weatherlore/4",
+        "Piloting/2",
+        "Shipwright/2"
+    ],
+    "Gladiator": [
+        "INITIATIVE/6",
+        "Sword/5",
+        "Shield/4",
+        "Spear/5",
+        "UNARMED/5",
+        "Physician/2"
+    ],
+    "Glassworker": [
+        "Glasswork/4",
+        "Ceramics/2",
+        "Alchemy/2",
+        "Script/5"
+    ],
+    "Harper/Skald": [
+        "SINGING/5",
+        "Folklore/4",
+        "ORATORY/4",
+        "Lute/4",
+        "Flute/3",
+        "Harp/3",
+        "Woodcraft/3",
+        "Script/5"
+    ],
+    "Herald": [
+        "Heraldry/4",
+        "ORATORY/4",
+        "RHETORIC/4",
+        "INTRIGUE/5",
+        "Drawing/4",
+        "Language/5",
+        "Script/5",
+        "INITIATIVE/6",
+        "Foraging/4",
+        "Survival/4",
+        "Physician/2",
+        "Weaponcraft/2",
+        "RIDING/6",
+        "Polearm(Lance)/6",
+        "Sword(Broadsword)/5",
+        "Axe(Handaxe)/5",
+        "Dagger/5",
+        "Shield(Knight)/6",
+        "Dancing/3"
+    ],
+    "Herdsman": [
+        "Animalcraft/4",
+        "Tracking/3",
+        "Survival/3",
+        "Weatherlore/4"
+    ],
+    "Hideworker": [
+        "Hidework/4",
+        "Textilecraft/2",
+        "Alchemy/2"
+    ],
+    "Hunter/Trapper": [
+        "Tracking/4",
+        "Bow/3",
+        "Dagger/4",
+        "Bow/4",
+        "STEALTH/4",
+        "Foraging/3",
+        "Survival/3",
+        "Hidework/2",
+        "Fletching/2"
+    ],
+    "Innkeeper": [
+        "Brewing/4",
+        "Cookery/4",
+        "RHETORIC/4",
+        "INTRIGUE/4",
+        "Language/3",
+        "Script/5"
+    ],
+    "Jeweler": [
+        "Jewelcraft/4",
+        "Metalcraft/3",
+        "Mining/2",
+        "Script/5"
+    ],
+    "Laborer/Longshoreman": [
+        "INTRIGUE/4",
+        "Seamanship/2"
+    ],
+    "Lady": [
+        "INTRIGUE/5",
+        "Folklore/4",
+        "Textilecraft(Needlework)/3",
+        "Dancing/3",
+        "Heraldry/3",
+        "Astrology/2",
+        "Drawing/2"
+    ],
+    "Lexigrapher": [
+        "Hidework(parchment&vellum)/4",
+        "Alchemy(Inkcraft)/4",
+        "Drawing/3",
+        "Script/5",
+        "Script/5"
+    ],
+    "Litigant": [
+        "Law/4",
+        "ORATORY/5",
+        "RHETORIC/5",
+        "INTRIGUE/5",
+        "Language/3",
+        "Script/5"
+    ],
+    "Locksmith": [
+        "Lockcraft/4",
+        "Metalcraft/3",
+        "Woodcraft/2",
+        "Script/5"
+    ],
+    "Mason": [
+        "Masonry/4",
+        "Woodcraft/3",
+        "Engineering/3",
+        "Mathematics/2",
+        "Script/5"
+    ],
+    "Mercantyler": [
+        "RHETORIC/5",
+        "INTRIGUE/5",
+        "Mathematics/3",
+        "Law/2",
+        "Dagger/2",
+        "Language/3",
+        "Script/5"
+    ],
+    "Metalsmith": [
+        "Metalcraft/4",
+        "Mining/2",
+        "Weaponcraft/2"
+    ],
+    "Miller/Millwright": [
+        "Milling/4",
+        "Engineering/3",
+        "Agriculture/3",
+        "Script/5"
+    ],
+    "Miner": [
+        "Mining/4",
+        "Engineering/3",
+        "Woodcraft/2",
+        "Metalcraft/2",
+        "Jewelcraft/1"
+    ],
+    "Ostler": [
+        "Animalcraft(Horsecraft)/4",
+        "RIDING/4",
+        "Hidework/3"
+    ],
+    "Perfumer": [
+        "Perfumery/4",
+        "Alchemy/3",
+        "Herblore/3",
+        "Embalming/2",
+        "Script/5"
+    ],
+    "Physician": [
+        "Physician/4",
+        "Herblore/3",
+        "Alchemy/3",
+        "Script/5"
+    ],
+    "Pilot": [
+        "Piloting/4",
+        "Weatherlore/4",
+        "Seamanship/3",
+        "Drawing/3",
+        "Bow/3",
+        "Shipwright/2",
+        "Mathematics/2",
+        "Script/5"
+    ],
+    "Potter": [
+        "Ceramics/4",
+        "Glasswork/2"
+    ],
+    "Prostitute/Pimp": [
+        "Lovecraft/4",
+        "INTRIGUE/4",
+        "Dagger/4"
+    ],
+    "Ratter": [
+        "Animalcraft(Rat)/5",
+        "Animalcraft(Dog)/4",
+        "Net/3",
+        "Club/5",
+        "Herblore/2"
+    ],
+    "Sage/Tutor": [
+        "Folklore/4",
+        "Mathematics/2",
+        "Language/3",
+        "Script/5",
+        "Script/5"
+    ],
+    "Salter": [
+        "Mining/3",
+        "Cookery/4",
+        "Survival/3",
+        "Fishing/3",
+        "Herblore/1"
+    ],
+    "Servant": [
+        "INTRIGUE/4",
+        "RHETORIC/4",
+        "Textilecraft/3"
+    ],
+    "Scribe": [
+        "Drawing/3",
+        "Mathematics/2",
+        "Language/3",
+        "Script/5",
+        "Script/5",
+        "Script/5"
+    ],
+    "Seaman": [
+        "Seamanship/4",
+        "CLIMBING/5",
+        "Club/5",
+        "Dagger/4",
+        "Fishing/3",
+        "Weatherlore/3",
+        "Piloting/2",
+        "Shipwright/2"
+    ],
+    "Shaman": [
+        "Ritual-Tribal/4",
+        "Folklore/4",
+        "INTRIGUE/4",
+        "MentalConflict/4",
+        "RHETORIC/4",
+        "Weatherlore/4",
+        "Dancing/3",
+        "Foraging/3",
+        "ORATORY/3",
+        "Embalming/2",
+        "Herblore/4",
+        "Physician/3",
+        "Survival/3",
+        "Tracking/2",
+        "Spear(Spear)/2"
+    ],
+    "Shipwright": [
+        "Shipwright/4",
+        "Woodcraft/3",
+        "Timbercraft/3",
+        "Metalcraft/2",
+        "Mathematics/2",
+        "Seamanship/2",
+        "Script/5"
+    ],
+    "Teamster": [
+        "RIDING/4",
+        "Animalcraft/3",
+        "Whip/3",
+        "Woodcraft/2"
+    ],
+    "Tentmaker": [
+        "Textilecraft/3",
+        "Hidework/3",
+        "Woodcraft/2"
+    ],
+    "Thatcher": [
+        "CLIMBING/5",
+        "Woodcraft/3",
+        "Engineering/2"
+    ],
+    "Thespian": [
+        "Acting/4",
+        "ORATORY/4",
+        "SINGING/4",
+        "Musician/3",
+        "Drawing/3"
+    ],
+    "Thief": [
+        "Legerdemain/4",
+        "AWARENESS/5",
+        "STEALTH/4",
+        "INTRIGUE/4",
+        "Lockcraft/3",
+        "Club/5",
+        "Dagger/4",
+        "Acrobatics/2"
+    ],
+    "Timberwright": [
+        "Timbercraft/4",
+        "Survival/3",
+        "Weatherlore/4",
+        "Woodcraft/3"
+    ],
+    "Toymaker": [
+        "Woodcraft/4",
+        "Lockcraft/3",
+        "Drawing/3",
+        "Hidework/2"
+    ],
+    "Weaponcrafter": [
+        "Weaponcraft/4",
+        "Sword/3",
+        "Bow/3",
+        "Metalcraft/3",
+        "Fletching/3",
+        "Mining/2"
+    ],
+    "Woodcrafter": [
+        "Woodcraft/5",
+        "Metalcraft/2",
+        "Hidework/2"
+    ],
+    "Cleric/Agrik": [
+        "Language(Azeryani)/3",
+        "Script(Zerin)",
+        "Ritual-Agrik/4",
+        "RHETORIC/4",
+        "INTRIGUE/4",
+        "MentalConflict/4",
+        "ORATORY/3",
+        "Folklore/3",
+        "Embalming/2",
+        "Physician/2",
+        "Law/2",
+        "Drawing/2",
+        "Script/5",
+        "INITIATIVE/6",
+        "UNARMED/4",
+        "Axe(Sickle)/4",
+        "Club(Mace)/4",
+        "Shield/4",
+        "Heraldry/3",
+        "Language(Surikal)/3"
+    ],
+    "Cleric/Halea": [
+        "Language(Karuia)/3",
+        "Script(Zerin)",
+        "Ritual-Halea/4",
+        "RHETORIC/4",
+        "INTRIGUE/4",
+        "MentalConflict/4",
+        "ORATORY/3",
+        "Folklore/3",
+        "Embalming/2",
+        "Physician/2",
+        "Law/2",
+        "Drawing/2",
+        "Heraldry/2",
+        "Script/5",
+        "Lovecraft/4",
+        "Dancing/3",
+        "Mathematics/2",
+        "Musician/2",
+        "Perfumery/2",
+        "Language/4",
+        "Script/5"
+    ],
+    "Cleric/Ilvir": [
+        "Language(Old Jarinese)/3",
+        "Script(Khruni)",
+        "Ritual-Ilvir/4",
+        "RHETORIC/4",
+        "INTRIGUE/4",
+        "MentalConflict/4",
+        "ORATORY/3",
+        "Folklore/3",
+        "Embalming/2",
+        "Law/2",
+        "Drawing/2",
+        "Heraldry/2",
+        "Script/5",
+        "Animalcraft(Ivashu)/4",
+        "Drawing/3",
+        "Physician/3",
+        "Tarotry/2",
+        "Herblore/2",
+        "Language(Ivashi)/3"
+    ],
+    "Cleric/Larani": [
+        "Language(Emela)/3",
+        "Script(Khruni)",
+        "Ritual-Larani/4",
+        "RHETORIC/4",
+        "INTRIGUE/4",
+        "MentalConflict/4",
+        "ORATORY/3",
+        "Folklore/3",
+        "Embalming/2",
+        "Physician/2",
+        "Law/2",
+        "Drawing/2",
+        "Script/5",
+        "INITIATIVE/6",
+        "UNARMED/4",
+        "Dagger/4",
+        "Sword/4",
+        "Shield/4",
+        "Heraldry/3"
+    ],
+    "Cleric/Morgath": [
+        "Language(Azeri)/3",
+        "Script(Nuvesarl)",
+        "Ritual-Morgath/4",
+        "RHETORIC/4",
+        "INTRIGUE/4",
+        "MentalConflict/4",
+        "ORATORY/3",
+        "Folklore/3",
+        "Physician/2",
+        "Law/2",
+        "Drawing/2",
+        "Heraldry/2",
+        "Script/5",
+        "Dagger/4",
+        "Embalming/3",
+        "Hidework(Human Skin)/2",
+        "Woodcraft(Masks)/2",
+        "Tarotry/2",
+        "Language(Ormauk)/3"
+    ],
+    "Cleric/Naveh": [
+        "Language(Besha)/3",
+        "Script(Neramic)",
+        "Ritual-Naveh/4",
+        "RHETORIC/4",
+        "INTRIGUE/4",
+        "MentalConflict/4",
+        "ORATORY/3",
+        "Folklore/3",
+        "Embalming/2",
+        "Physician/2",
+        "Law/2",
+        "Drawing/2",
+        "Heraldry/2",
+        "Script/5",
+        "INITIATIVE/6",
+        "UNARMED/5",
+        "STEALTH/5",
+        "Dagger/4",
+        "Acrobatics/3",
+        "Legerdemain/2",
+        "Lockcraft/2"
+    ],
+    "Cleric/Peoni": [
+        "Language(Emela)/3",
+        "Script(Khruni)",
+        "Ritual-Peoni/4",
+        "RHETORIC/4",
+        "INTRIGUE/4",
+        "MentalConflict/4",
+        "ORATORY/3",
+        "Folklore/3",
+        "Embalming/2",
+        "Law/2",
+        "Drawing/2",
+        "Heraldry/2",
+        "Script/5",
+        "Agriculture/4",
+        "Animalcraft/3",
+        "Herblore/3",
+        "Physician/3",
+        "Weatherlore/3",
+        "Textilecraft/2"
+    ],
+    "Cleric/Sarajin": [
+        "Language(Ivinian)/3",
+        "Script(Runic)",
+        "Ritual-Sarajin/4",
+        "RHETORIC/4",
+        "INTRIGUE/4",
+        "MentalConflict/4",
+        "ORATORY/3",
+        "Folklore/3",
+        "Embalming/2",
+        "Physician/2",
+        "Law/2",
+        "Drawing/2",
+        "Heraldry/2",
+        "Script/5",
+        "INITIATIVE/6",
+        "UNARMED/4",
+        "Axe/4",
+        "Dagger/4",
+        "Shield/4",
+        "Runecraft/3",
+        "Seamanship/2"
+    ],
+    "Cleric/Save-K'nor": [
+        "Language(Azeri)/3",
+        "Script(Tianta)",
+        "Ritual-Save-K'nor/4",
+        "RHETORIC/4",
+        "INTRIGUE/4",
+        "MentalConflict/4",
+        "ORATORY/3",
+        "Folklore/3",
+        "Embalming/2",
+        "Physician/2",
+        "Law/2",
+        "Drawing/2",
+        "Heraldry/2",
+        "Script/5",
+        "Drawing/3",
+        "Mathematics/3",
+        "Law/3",
+        "Heraldry/2",
+        "Language/4",
+        "Language/3",
+        "Script/5"
+    ],
+    "Cleric/Siem": [
+        "Language(Sindarin)/3",
+        "Script(Selenian)",
+        "Ritual-Siem/4",
+        "RHETORIC/4",
+        "INTRIGUE/4",
+        "MentalConflict/4",
+        "ORATORY/3",
+        "Folklore/3",
+        "Embalming/2",
+        "Physician/2",
+        "Law/2",
+        "Drawing/2",
+        "Heraldry/2",
+        "Script/5",
+        "Astrology/4",
+        "Runecraft/3",
+        "Foraging/3",
+        "Survival/3",
+        "Weatherlore/3",
+        "Tarotry/2",
+        "Musician/2",
+        "Herblore/2"
+    ],
+    "Feudal Militia": [
+        "INITIATIVE/5",
+        "Foraging/4",
+        "Survival/4",
+        "Heraldry/2",
+        "Physician/2",
+        "Weaponcraft/2",
+        "Spear/4",
+        "Shield(Round)/4"
+    ],
+    "Feudal Yeoman, Foot": [
+        "INITIATIVE/5",
+        "Foraging/4",
+        "Survival/4",
+        "Heraldry/2",
+        "Physician/2",
+        "Weaponcraft/2",
+        "Spear/5",
+        "Sword(Shortsword)/4",
+        "Dagger/4",
+        "Shield(Round)/4"
+    ],
+    "Feudal Guardsman, Light": [
+        "INITIATIVE/5",
+        "Foraging/4",
+        "Survival/4",
+        "Heraldry/2",
+        "Physician/2",
+        "Weaponcraft/2",
+        "Spear/6",
+        "Sword(Shortsword)/5",
+        "Dagger/5",
+        "Shield(Round)/5"
+    ],
+    "Feudal Guardsman, Heavy": [
+        "INITIATIVE/5",
+        "Foraging/4",
+        "Survival/4",
+        "Heraldry/2",
+        "Physician/2",
+        "Weaponcraft/2",
+        "Spear/6",
+        "Sword(Falchion)/5",
+        "Dagger/5",
+        "Shield(Round)/5"
+    ],
+    "Feudal Yeoman, Shortbow": [
+        "INITIATIVE/5",
+        "Foraging/4",
+        "Survival/4",
+        "Heraldry/2",
+        "Physician/2",
+        "Weaponcraft/2",
+        "Bow(Shortbow)/5",
+        "Sword(Falchion)/5",
+        "Dagger/5",
+        "Shield(Buckler)/5"
+    ],
+    "Feudal Yeoman, Longbow": [
+        "INITIATIVE/5",
+        "Foraging/4",
+        "Survival/4",
+        "Heraldry/2",
+        "Physician/2",
+        "Weaponcraft/2",
+        "Bow(Longbow)/5",
+        "Sword(Falchion)/5",
+        "Dagger/5",
+        "Shield(Buckler)/5"
+    ],
+    "Feudal Knight, Medium": [
+        "INITIATIVE/6",
+        "Foraging/4",
+        "Survival/4",
+        "Heraldry/2",
+        "Physician/2",
+        "Weaponcraft/2",
+        "RIDING/6",
+        "Polearm(Lance)/6",
+        "Sword(Broadsword)/5",
+        "Axe(Handaxe)/5",
+        "Dagger/5",
+        "Shield(Knight)/6",
+        "Dancing/3"
+    ],
+    "Feudal Knight, Heavy": [
+        "INITIATIVE/6",
+        "Foraging/4",
+        "Survival/4",
+        "Heraldry/2",
+        "Physician/2",
+        "Weaponcraft/2",
+        "RIDING/6",
+        "Polearm(Lance)/6",
+        "Sword(Bastard Sword)/5",
+        "Axe(Handaxe)/5",
+        "Dagger/5",
+        "Shield(Kite)/6",
+        "Dancing/3"
+    ],
+    "Order Infantry, Light": [
+        "INITIATIVE/5",
+        "Foraging/4",
+        "Survival/4",
+        "Heraldry/2",
+        "Physician/2",
+        "Weaponcraft/2",
+        "Spear/6",
+        "Sword(Shortsword)/5",
+        "Dagger/5",
+        "Shield(Round)/5"
+    ],
+    "Order Infantry, Heavy": [
+        "INITIATIVE/5",
+        "Foraging/4",
+        "Survival/4",
+        "Heraldry/2",
+        "Physician/2",
+        "Weaponcraft/2",
+        "Spear/6",
+        "Sword(Falchion)/5",
+        "Dagger/5",
+        "Shield(Round)/5"
+    ],
+    "Order Archer": [
+        "INITIATIVE/5",
+        "Foraging/4",
+        "Survival/4",
+        "Heraldry/2",
+        "Physician/2",
+        "Weaponcraft/2",
+        "Bow(Shortbow)/5",
+        "Sword(Shortsword)/5",
+        "Dagger/5",
+        "Shield(Buckler)/5"
+    ],
+    "Order Knight, Medium": [
+        "INITIATIVE/6",
+        "Foraging/4",
+        "Survival/4",
+        "Heraldry/2",
+        "Physician/2",
+        "Weaponcraft/2",
+        "RIDING/6",
+        "Polearm(Lance)/6",
+        "Sword(Broadsword)/5",
+        "Club(Mace)/5",
+        "Dagger/5",
+        "Shield(Knight)/6",
+        "Dancing/3"
+    ],
+    "Order Knight, Heavy": [
+        "INITIATIVE/6",
+        "Foraging/4",
+        "Survival/4",
+        "Heraldry/2",
+        "Physician/2",
+        "Weaponcraft/2",
+        "RIDING/6",
+        "Polearm(Lance)/6",
+        "Sword(Bastard Sword)/5",
+        "Club(Mace)/5",
+        "Dagger/5",
+        "Shield(Kite)/6",
+        "Dancing/3"
+    ],
+    "Imperial Militia": [
+        "INITIATIVE/5",
+        "Foraging/4",
+        "Survival/4",
+        "Heraldry/2",
+        "Physician/2",
+        "Weaponcraft/2",
+        "Spear/4",
+        "Shield(Round)/4"
+    ],
+    "Imperial Legionnaire, Light": [
+        "INITIATIVE/5",
+        "Foraging/4",
+        "Survival/4",
+        "Heraldry/2",
+        "Physician/2",
+        "Weaponcraft/2",
+        "Spear/5",
+        "Sword(Shortsword)/4",
+        "Dagger/4",
+        "Shield(Tower)/5"
+    ],
+    "Imperial Legionnaire, Medium": [
+        "INITIATIVE/5",
+        "Foraging/4",
+        "Survival/4",
+        "Heraldry/2",
+        "Physician/2",
+        "Weaponcraft/2",
+        "Spear/6",
+        "Sword(Shortsword)/5",
+        "Dagger/4",
+        "Shield(Tower)/5"
+    ],
+    "Imperial Legionnaire, Shortbow": [
+        "INITIATIVE/5",
+        "Foraging/4",
+        "Survival/4",
+        "Heraldry/2",
+        "Physician/2",
+        "Weaponcraft/2",
+        "Bow(Shortbow)/6",
+        "Sword(Shortsword)/5",
+        "Dagger/5",
+        "Shield(Buckler)/5"
+    ],
+    "Patrician": [
+        "INITIATIVE/6",
+        "Foraging/4",
+        "Survival/4",
+        "Heraldry/2",
+        "Physician/2",
+        "Weaponcraft/2",
+        "RIDING/6",
+        "Polearm(Lance)/6",
+        "Sword(Broadsword)/5",
+        "Axe(Handaxe)/5",
+        "Dagger/5",
+        "Shield(Kite)/6",
+        "Dancing/3"
+    ],
+    "Viking Clansman, Unarmored": [
+        "INITIATIVE/5",
+        "Foraging/4",
+        "Survival/4",
+        "Heraldry/2",
+        "Physician/2",
+        "Weaponcraft/2",
+        "Spear/5",
+        "Shield(Round)/5",
+        "Dagger(Keltan)/4",
+        "Seamanship/3"
+    ],
+    "Viking Clansman, Light": [
+        "INITIATIVE/5",
+        "Foraging/4",
+        "Survival/4",
+        "Heraldry/2",
+        "Physician/2",
+        "Weaponcraft/2",
+        "Spear/6",
+        "Shield(Round)/6",
+        "Bow((Shortbow))/5",
+        "Dagger(Keltan)/5",
+        "Seamanship/3"
+    ],
+    "Viking Huscarl, Medium Foot": [
+        "INITIATIVE/6",
+        "Foraging/4",
+        "Survival/4",
+        "Heraldry/2",
+        "Physician/2",
+        "Weaponcraft/2",
+        "Axe(Battleaxe)/6",
+        "Sword(Broadsword)/5",
+        "Dagger(Keltan)/5",
+        "Shield(Round)/5",
+        "Seamanship/3",
+        "Piloting/2"
+    ],
+    "Viking Huscarl, Light Horse": [
+        "INITIATIVE/6",
+        "Foraging/4",
+        "Survival/4",
+        "Heraldry/2",
+        "Physician/2",
+        "Weaponcraft/2",
+        "RIDING/6",
+        "Polearm(Lance)/6",
+        "Sword(Broadsword)/5",
+        "Shield(Round)/5",
+        "Seamanship/3",
+        "Piloting/2"
+    ],
+    "Khuzdul Clansman": [
+        "INITIATIVE/5",
+        "Foraging/4",
+        "Survival/4",
+        "Heraldry/2",
+        "Physician/2",
+        "Weaponcraft/2",
+        "Spear/6",
+        "Axe(Handaxe)/5",
+        "Shield(Round)/5"
+    ],
+    "Khuzdul Low Guard": [
+        "INITIATIVE/5",
+        "Foraging/4",
+        "Survival/4",
+        "Heraldry/2",
+        "Physician/2",
+        "Weaponcraft/2",
+        "Spear/6",
+        "Axe(Battleaxe)/5",
+        "Dagger/5",
+        "Shield(Round)/5"
+    ],
+    "Khuzdul High Guard": [
+        "INITIATIVE/6",
+        "Foraging/4",
+        "Survival/4",
+        "Heraldry/2",
+        "Physician/2",
+        "Weaponcraft/2",
+        "Polearm(Poleaxe)/6",
+        "Club(Mace)/5",
+        "Dagger/5",
+        "Shield(Round)/5"
+    ],
+    "Sindarin Ranger, Unarmored": [
+        "INITIATIVE/5",
+        "Foraging/4",
+        "Survival/4",
+        "Heraldry/2",
+        "Physician/2",
+        "Weaponcraft/2",
+        "Bow(Hartbow)/6",
+        "Sword(Longknife)/5",
+        "Dagger/5",
+        "Shield(Buckler)/5"
+    ],
+    "Sindarin Ranger, Light": [
+        "INITIATIVE/5",
+        "Foraging/4",
+        "Survival/4",
+        "Heraldry/2",
+        "Physician/2",
+        "Weaponcraft/2",
+        "Bow(Hartbow)/6",
+        "Sword(Longknife)/5",
+        "Dagger/5",
+        "Shield(Buckler)/5"
+    ],
+    "Sindarin Guardian": [
+        "INITIATIVE/5",
+        "Foraging/4",
+        "Survival/4",
+        "Heraldry/2",
+        "Physician/2",
+        "Weaponcraft/2",
+        "Spear/7",
+        "Sword(Longknife)/6",
+        "Dagger/5",
+        "Shield(Round)/6"
+    ],
+    "Sindarin Horsebow": [
+        "INITIATIVE/5",
+        "Foraging/4",
+        "Survival/4",
+        "Heraldry/2",
+        "Physician/2",
+        "Weaponcraft/2",
+        "RIDING/6",
+        "Bow(Hartbow)/6",
+        "Polearm(Lance)/5",
+        "Sword(Longknife)/5",
+        "Shield(Knight)/6",
+        "Dancing/3",
+        "Musician/2"
+    ],
+    "Sindarin Knight": [
+        "INITIATIVE/6",
+        "Foraging/4",
+        "Survival/4",
+        "Heraldry/2",
+        "Physician/2",
+        "Weaponcraft/2",
+        "RIDING/7",
+        "Polearm(Lance)/7",
+        "Sword(Longknife)/6",
+        "Shield(Knight)/6",
+        "Dancing/3",
+        "Musician/2"
+    ],
+    "Shek-Pvar/Lyahvi*": [
+        "Lyahvi/2",
+        "Neutral/1",
+        "Script/2",
+        "Folklore/4",
+        "Mathematics/2",
+        "Legerdemain/3",
+        "Glasswork/3",
+        "Jewelcraft/3"
+    ],
+    "Shek-Pvar/Lyahvi**": [
+        "Lyahvi/2",
+        "Neutral/2",
+        "Script/2",
+        "Folklore/4",
+        "Mathematics/2",
+        "Legerdemain/3",
+        "Glasswork/3",
+        "Jewelcraft/3"
+    ],
+    "Shek-Pvar/Lyahvi***": [
+        "Lyahvi/3",
+        "Neutral/2",
+        "Script/2",
+        "Folklore/4",
+        "Mathematics/2",
+        "Legerdemain/3",
+        "Glasswork/3",
+        "Jewelcraft/3"
+    ],
+    "Shek-Pvar/Lyahvi****": [
+        "Lyahvi/3",
+        "Neutral/3",
+        "Script/2",
+        "Folklore/4",
+        "Mathematics/2",
+        "Legerdemain/3",
+        "Glasswork/3",
+        "Jewelcraft/3"
+    ],
+    "Shek-Pvar/Lyahvi*****": [
+        "Lyahvi/4",
+        "Neutral/3",
+        "Script/2",
+        "Folklore/4",
+        "Mathematics/2",
+        "Legerdemain/3",
+        "Glasswork/3",
+        "Jewelcraft/3"
+    ],
+    "Shek-Pvar/Peleahn*": [
+        "Peleahn/2",
+        "Neutral/1",
+        "Script/2",
+        "Folklore/4",
+        "Mathematics/2",
+        "Alchemy/3",
+        "Cookery/4",
+        "Metalcraft/1",
+        "Weaponcraft/1"
+    ],
+    "Shek-Pvar/Peleahn**": [
+        "Peleahn/2",
+        "Neutral/2",
+        "Script/2",
+        "Folklore/4",
+        "Mathematics/2",
+        "Alchemy/3",
+        "Cookery/4",
+        "Metalcraft/1",
+        "Weaponcraft/1"
+    ],
+    "Shek-Pvar/Peleahn***": [
+        "Peleahn/3",
+        "Neutral/2",
+        "Script/2",
+        "Folklore/4",
+        "Mathematics/2",
+        "Alchemy/3",
+        "Cookery/4",
+        "Metalcraft/1",
+        "Weaponcraft/1"
+    ],
+    "Shek-Pvar/Peleahn****": [
+        "Peleahn/3",
+        "Neutral/3",
+        "Script/2",
+        "Folklore/4",
+        "Mathematics/2",
+        "Alchemy/3",
+        "Cookery/4",
+        "Metalcraft/1",
+        "Weaponcraft/1"
+    ],
+    "Shek-Pvar/Peleahn*****": [
+        "Peleahn/4",
+        "Neutral/3",
+        "Script/2",
+        "Folklore/4",
+        "Mathematics/2",
+        "Alchemy/3",
+        "Cookery/4",
+        "Metalcraft/1",
+        "Weaponcraft/1"
+    ],
+    "Shek-Pvar/Jmorvi*": [
+        "Jmorvi/2",
+        "Neutral/1",
+        "Script/2",
+        "Folklore/4",
+        "Mathematics/2",
+        "Metalcraft/3",
+        "Lockcraft/3",
+        "Mining(Mineralogy)/2",
+        "Weaponcraft/2"
+    ],
+    "Shek-Pvar/Jmorvi**": [
+        "Jmorvi/2",
+        "Neutral/2",
+        "Script/2",
+        "Folklore/4",
+        "Mathematics/2",
+        "Metalcraft/3",
+        "Lockcraft/3",
+        "Mining(Mineralogy)/2",
+        "Weaponcraft/2"
+    ],
+    "Shek-Pvar/Jmorvi***": [
+        "Jmorvi/3",
+        "Neutral/2",
+        "Script/2",
+        "Folklore/4",
+        "Mathematics/2",
+        "Metalcraft/3",
+        "Lockcraft/3",
+        "Mining(Mineralogy)/2",
+        "Weaponcraft/2"
+    ],
+    "Shek-Pvar/Jmorvi****": [
+        "Jmorvi/3",
+        "Neutral/3",
+        "Script/2",
+        "Folklore/4",
+        "Mathematics/2",
+        "Metalcraft/3",
+        "Lockcraft/3",
+        "Mining(Mineralogy)/2",
+        "Weaponcraft/2"
+    ],
+    "Shek-Pvar/Jmorvi*****": [
+        "Jmorvi/4",
+        "Neutral/3",
+        "Script/2",
+        "Folklore/4",
+        "Mathematics/2",
+        "Metalcraft/3",
+        "Lockcraft/3",
+        "Mining(Mineralogy)/2",
+        "Weaponcraft/2"
+    ],
+    "Shek-Pvar/Fyvria*": [
+        "Fyvria/2",
+        "Neutral/1",
+        "Script/2",
+        "Folklore/4",
+        "Mathematics/2",
+        "Herblore/3",
+        "Agriculture/3",
+        "Animalcraft/3",
+        "Embalming/2",
+        "Foraging/3",
+        "Physician/2",
+        "Survival/3",
+        "Tracking/2"
+    ],
+    "Shek-Pvar/Fyvria**": [
+        "Fyvria/2",
+        "Neutral/2",
+        "Script/2",
+        "Folklore/4",
+        "Mathematics/2",
+        "Herblore/3",
+        "Agriculture/3",
+        "Animalcraft/3",
+        "Embalming/2",
+        "Foraging/3",
+        "Physician/2",
+        "Survival/3",
+        "Tracking/2"
+    ],
+    "Shek-Pvar/Fyvria***": [
+        "Fyvria/3",
+        "Neutral/2",
+        "Script/2",
+        "Folklore/4",
+        "Mathematics/2",
+        "Herblore/3",
+        "Agriculture/3",
+        "Animalcraft/3",
+        "Embalming/2",
+        "Foraging/3",
+        "Physician/2",
+        "Survival/3",
+        "Tracking/2"
+    ],
+    "Shek-Pvar/Fyvria****": [
+        "Fyvria/3",
+        "Neutral/3",
+        "Script/2",
+        "Folklore/4",
+        "Mathematics/2",
+        "Herblore/3",
+        "Agriculture/3",
+        "Animalcraft/3",
+        "Embalming/2",
+        "Foraging/3",
+        "Physician/2",
+        "Survival/3",
+        "Tracking/2"
+    ],
+    "Shek-Pvar/Fyvria*****": [
+        "Fyvria/4",
+        "Neutral/3",
+        "Script/2",
+        "Folklore/4",
+        "Mathematics/2",
+        "Herblore/3",
+        "Agriculture/3",
+        "Animalcraft/3",
+        "Embalming/2",
+        "Foraging/3",
+        "Physician/2",
+        "Survival/3",
+        "Tracking/2"
+    ],
+    "Shek-Pvar/Odivshe*": [
+        "Odivshe/2",
+        "Neutral/1",
+        "Script/2",
+        "Folklore/4",
+        "Mathematics/2",
+        "Swimming/4",
+        "Fishing/3",
+        "Piloting/2",
+        "Seamanship/2",
+        "Brewing/2"
+    ],
+    "Shek-Pvar/Odivshe**": [
+        "Odivshe/2",
+        "Neutral/2",
+        "Script/2",
+        "Folklore/4",
+        "Mathematics/2",
+        "Swimming/4",
+        "Fishing/3",
+        "Piloting/2",
+        "Seamanship/2",
+        "Brewing/2"
+    ],
+    "Shek-Pvar/Odivshe***": [
+        "Odivshe/3",
+        "Neutral/2",
+        "Script/2",
+        "Folklore/4",
+        "Mathematics/2",
+        "Swimming/4",
+        "Fishing/3",
+        "Piloting/2",
+        "Seamanship/2",
+        "Brewing/2"
+    ],
+    "Shek-Pvar/Odivshe****": [
+        "Odivshe/3",
+        "Neutral/3",
+        "Script/2",
+        "Folklore/4",
+        "Mathematics/2",
+        "Swimming/4",
+        "Fishing/3",
+        "Piloting/2",
+        "Seamanship/2",
+        "Brewing/2"
+    ],
+    "Shek-Pvar/Odivshe*****": [
+        "Odivshe/4",
+        "Neutral/3",
+        "Script/2",
+        "Folklore/4",
+        "Mathematics/2",
+        "Swimming/4",
+        "Fishing/3",
+        "Piloting/2",
+        "Seamanship/2",
+        "Brewing/2"
+    ],
+    "Shek-Pvar/Savorya*": [
+        "Savorya/2",
+        "Neutral/1",
+        "Script/2",
+        "Folklore/4",
+        "Mathematics/2",
+        "MentalConflict/5",
+        "Drawing/4",
+        "Runecraft/3",
+        "Tarotry/3"
+    ],
+    "Shek-Pvar/Savorya**": [
+        "Savorya/2",
+        "Neutral/2",
+        "Script/2",
+        "Folklore/4",
+        "Mathematics/2",
+        "MentalConflict/5",
+        "Drawing/4",
+        "Runecraft/3",
+        "Tarotry/3"
+    ],
+    "Shek-Pvar/Savorya***": [
+        "Savorya/3",
+        "Neutral/2",
+        "Script/2",
+        "Folklore/4",
+        "Mathematics/2",
+        "MentalConflict/5",
+        "Drawing/4",
+        "Runecraft/3",
+        "Tarotry/3"
+    ],
+    "Shek-Pvar/Savorya****": [
+        "Savorya/3",
+        "Neutral/3",
+        "Script/2",
+        "Folklore/4",
+        "Mathematics/2",
+        "MentalConflict/5",
+        "Drawing/4",
+        "Runecraft/3",
+        "Tarotry/3"
+    ],
+    "Shek-Pvar/Savorya*****": [
+        "Savorya/4",
+        "Neutral/3",
+        "Script/2",
+        "Folklore/4",
+        "Mathematics/2",
+        "MentalConflict/5",
+        "Drawing/4",
+        "Runecraft/3",
+        "Tarotry/3"
+    ]
+}
+
+////////////////////////////////////////////////////////////
+
+
+///////////////////////////////////
+
+
+
+
+
+
+
+///////////////////////////////
+
+
+var occupation_time = {
+    "Alchemist": "7",
+    "Animal Trainer": "3",
+    "Apothecary": "6",
+    "Astrologer": "4",
+    "Baliff": "8",
+    "Beggar": "1",
+    "Cartographer/Artist": "5",
+    "Chandler": "4",
+    "Charcoaler": "3",
+    "Chieftan": "15",
+    "Clothier": "6",
+    "Cook/Servant": "3",
+    "Courtesan": "3",
+    "Embalmer": "7",
+    "Farmer": "4",
+    "Fisherman": "5",
+    "Gladiator": "4",
+    "Glassworker": "7",
+    "Harper/Skald": "7",
+    "Herald": "12",
+    "Herdsman": "2",
+    "Hideworker": "5",
+    "Hunter/Trapper": "5",
+    "Innkeeper": "5",
+    "Jeweler": "7",
+    "Laborer/Longshoreman": "1",
+    "Lady": "7",
+    "Lexigrapher": "5",
+    "Litigant": "7",
+    "Locksmith": "6",
+    "Mason": "7",
+    "Mercantyler": "6",
+    "Metalsmith": "6",
+    "Miller/Millwright": "6",
+    "Miner": "6",
+    "Ostler": "5",
+    "Perfumer": "6",
+    "Physician": "7",
+    "Pilot": "9",
+    "Potter": "6",
+    "Prostitute/Pimp": "1",
+    "Ratter": "3",
+    "Sage/Tutor": "5",
+    "Salter": "4",
+    "Servant": "1",
+    "Scribe": "5",
+    "Seaman": "5",
+    "Shaman": "8",
+    "Shipwright": "7",
+    "Teamster": "2",
+    "Tentmaker": "5",
+    "Thatcher": "4",
+    "Thespian": "7",
+    "Thief": "3",
+    "Timberwright": "6",
+    "Toymaker": "5",
+    "Weaponcrafter": "7",
+    "Woodcrafter": "7",
+    "Cleric/Agrik": "9",
+    "Cleric/Halea": "9",
+    "Cleric/Ilvir": "9",
+    "Cleric/Larani": "9",
+    "Cleric/Morgath": "9",
+    "Cleric/Naveh": "9",
+    "Cleric/Peoni": "9",
+    "Cleric/Sarajin": "9",
+    "Cleric/Save-K'nor": "9",
+    "Cleric/Siem": "9",
+    "Feudal Militia": "4",
+    "Feudal Yeoman, Foot": "5",
+    "Feudal Guardsman, Light": "5",
+    "Feudal Guardsman, Heavy": "5",
+    "Feudal Yeoman, Shortbow": "6",
+    "Feudal Yeoman, Longbow": "6",
+    "Feudal Knight, Medium": "7",
+    "Feudal Knight, Heavy": "7",
+    "Order Infantry, Light": "5",
+    "Order Infantry, Heavy": "5",
+    "Order Archer": "6",
+    "Order Knight, Medium": "7",
+    "Order Knight, Heavy": "7",
+    "Imperial Militia": "5",
+    "Imperial Legionnaire, Light": "5",
+    "Imperial Legionnaire, Medium": "5",
+    "Imperial Legionnaire, Shortbow": "6",
+    "Patrician": "7",
+    "Viking Clansman, Unarmored": "4",
+    "Viking Clansman, Light": "5",
+    "Viking Huscarl, Medium Foot": "6",
+    "Viking Huscarl, Light Horse": "6",
+    "Khuzdul Clansman": "15",
+    "Khuzdul Low Guard": "20",
+    "Khuzdul High Guard": "30",
+    "Sindarin Ranger, Unarmored": "10",
+    "Sindarin Ranger, Light": "11",
+    "Sindarin Guardian": "12",
+    "Sindarin Horsebow": "15",
+    "Sindarin Knight": "20",
+    "Shek-Pvar/Lyahvi*": "7",
+    "Shek-Pvar/Lyahvi**": "7",
+    "Shek-Pvar/Lyahvi***": "7",
+    "Shek-Pvar/Lyahvi****": "7",
+    "Shek-Pvar/Lyahvi*****": "7",
+    "Shek-Pvar/Peleahn*": "7",
+    "Shek-Pvar/Peleahn**": "7",
+    "Shek-Pvar/Peleahn***": "7",
+    "Shek-Pvar/Peleahn****": "7",
+    "Shek-Pvar/Peleahn*****": "7",
+    "Shek-Pvar/Jmorvi*": "7",
+    "Shek-Pvar/Jmorvi**": "7",
+    "Shek-Pvar/Jmorvi***": "7",
+    "Shek-Pvar/Jmorvi****": "7",
+    "Shek-Pvar/Jmorvi*****": "7",
+    "Shek-Pvar/Fyvria*": "7",
+    "Shek-Pvar/Fyvria**": "7",
+    "Shek-Pvar/Fyvria***": "7",
+    "Shek-Pvar/Fyvria****": "7",
+    "Shek-Pvar/Fyvria*****": "7",
+    "Shek-Pvar/Odivshe*": "7",
+    "Shek-Pvar/Odivshe**": "7",
+    "Shek-Pvar/Odivshe***": "7",
+    "Shek-Pvar/Odivshe****": "7",
+    "Shek-Pvar/Odivshe*****": "7",
+    "Shek-Pvar/Savorya*": "7",
+    "Shek-Pvar/Savorya**": "7",
+    "Shek-Pvar/Savorya***": "7",
+    "Shek-Pvar/Savorya****": "7",
+    "Shek-Pvar/Savorya*****": "7"
+}
+
